Added a favorites controller
