+++ app/controllers/islands_controller.rb
@@ -1,5 +1,5 @@
 class IslandsController < ApplicationController
-  before_action :authenticate_user!, only: [:favorite, :unfavorite]
+  before_action :authenticate_user!, only: [:create]
 
   def index
     @islands = Island.all
@@ -12,22 +12,6 @@ class IslandsController < ApplicationController
   def new
     @island = Island.new
     @my_islands = current_user.islands
-  end
-
-  def favorite
-    @island = Island.find(params[:id])
-    current_user.favorite_islands << @island unless current_user.favorite_islands.include?(@island)
-    redirect_to root_path, notice: "#{@island.name} has been added to your favorites."
-  end
-
-  def unfavorite
-    @island = Island.find(params[:id])
-    current_user.favorite_islands.delete(@island)
-    redirect_to root_path, notice: "#{@island.name} has been removed from your favorites."
-  end
-
-  def favorites
-    @favorite_islands = current_user.favorite_islands
   end
 
   def create
