+++ app/models/island.rb
@@ -1,5 +1,5 @@
 class Island < ApplicationRecord
   belongs_to :user
   has_many :bookings, dependent: :destroy
-  has_many :favorites
+  has_many :favorites, dependent: :destroy
 end
