+++ app/models/user.rb
@@ -2,7 +2,6 @@ class User < ApplicationRecord
   has_many :islands
   has_many :bookings
   has_many :favorites
-  has_many :favorite_islands, through: :favorites, source: :island
   # Include default devise modules. Others available are:
   # :confirmable, :lockable, :timeoutable, :trackable and :omniauthable
   devise :database_authenticatable, :registerable,
