+++ app/views/islands/new.html.erb
@@ -1,4 +1,4 @@
-<h1>My Islands</h1>
+<%# <h1>My Islands</h1>
 <%# <% if @my_islands.any? %>
   <ul>
     <%# <%= @my_islands.each do |island| %>
@@ -6,7 +6,7 @@
   </ul>
 <%# <% else %>
   <%# <p><%= "You have not created any islands yet." %></p> %> %>
-<%# <% end %> %> %>
+<% end %> %> %>
 
 
 <h2>Create a New Island</h2>
