+++ app/views/shared/_navbar.html.erb
@@ -22,7 +22,7 @@
               <%= link_to "View Profile", "#", class: "dropdown-item" %>
               <%= link_to "Edit Profile", edit_user_registration_path, class: "dropdown-item" %>
               <%= link_to "My islands", "islands/new", class: "dropdown-item" %>
-              <%= link_to "My favorites", "/favorites", class: "dropdown-item" %>
+              <%= link_to "My favorites", favorites_path, class: "dropdown-item" %>
               <%= link_to "My Bookings", bookings_path, class: "dropdown-item" %>
               <%= link_to "Log out", destroy_user_session_path, data: {turbo_method: :delete}, class: "dropdown-item" %>
             </div>
